+++ xApplication.h
@@ -142,6 +142,12 @@ private slots:
      */
     void pulseAudioError(const QString& errorMsg);
     /**
+     * Show dialog if an error occurs trying to connect to the movie player (vlc).
+     *
+     * @param errorMsg the error message as string.
+     */
+    void moviePlayerError(const QString& errorMsg);
+    /**
      * Show dialog for unknown entries in the music table of the database.
      *
      * @param entries a list of tuples of artist, album and track unknown.