+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 
 ## Unreleased
 
+- Add full window mode for music visualization.
 - Fix default audio/subtitle setting for movie files.
 - Fix crashes when switching libraries without properly configured music libraries.
 - Enable reindex menu entry only if BluOS player library is used.
 - Improve switch times in movie view when scanning movie files.
+- Improve error handling for movie player.
 - Disconnect Rotel amp controls when closing.